src/python/requirements.txt
tqdm
h5py
numpy
tensorflow-gpu==2.2
